% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast-walk.R
\name{ast_depth}
\alias{ast_depth}
\title{Compute maximum AST nesting depth}
\usage{
ast_depth(expr, depth = 0L)
}
\arguments{
\item{expr}{A language object.}

\item{depth}{Integer. Current depth (internal).}
}
\value{
Integer. Maximum nesting depth.
}
\description{
Compute maximum AST nesting depth
}
\examples{
expr <- parse_code("f(g(h(1)))")[[1]]
ast_depth(expr)

expr2 <- parse_code("x <- 1")[[1]]
ast_depth(expr2)
}
