% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output-pii.R
\name{output_to_text}
\alias{output_to_text}
\title{Convert an R object to scannable text}
\usage{
output_to_text(x)
}
\arguments{
\item{x}{An arbitrary R object.}
}
\value{
Character(1). A text representation of \code{x}.
}
\description{
Converts arbitrary R objects to a single character string for pattern
scanning by output guardrails.
}
\examples{
output_to_text("hello")
output_to_text(data.frame(a = 1:3, b = letters[1:3]))
output_to_text(list(x = 1, y = "two"))
}
\keyword{internal}
