% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output-pii.R
\name{output_to_text}
\alias{output_to_text}
\title{Convert an R object to scannable text}
\usage{
output_to_text(x)
}
\arguments{
\item{x}{An arbitrary R object.}
}
\value{
Character(1). A text representation of \code{x}.
}
\description{
Converts arbitrary R objects to a single character string for pattern
scanning by output guardrails.
}
\keyword{internal}
